refactor: extract override logic
